+++ src/main.js
@@ -286,21 +286,24 @@ class MiniGraphCard extends LitElement {
   }
 
   renderState(entityConfig, id) {
-    const isPrimary = id === 0;
+    const isPrimary = id === 0; // rendering main state element?
     if (isPrimary || entityConfig.show_state) {
-      const { entity, value: tooltipValue } = this.tooltip;
       const state = this.getEntityState(id);
+      // use tooltip data for main state element, if tooltip is active
+      const { entity: tooltipEntity, value: tooltipValue } = this.tooltip;
+      const value = isPrimary && tooltipEntity !== undefined ? tooltipValue : state;
+      const entity = isPrimary && tooltipEntity !== undefined ? tooltipEntity : id;
       return html`
         <div
           class="state ${!isPrimary && 'state--small'}"
           @click=${e => this.handlePopup(e, this.entity[id])}
-          style=${entityConfig.state_adaptive_color ? `color: ${this.computeColor(state, id)};` : ''}>
-          ${entityConfig.show_indicator ? this.renderIndicator(state, id) : ''}
+          style=${entityConfig.state_adaptive_color ? `color: ${this.computeColor(value, entity)}` : ''}>
+          ${entityConfig.show_indicator ? this.renderIndicator(value, entity) : ''}
           <span class="state__value ellipsis">
-            ${this.computeState((isPrimary && tooltipValue !== undefined) ? tooltipValue : state)}
+            ${this.computeState(value)}
           </span>
           <span class="state__uom ellipsis">
-            ${this.computeUom(isPrimary && entity || id)}
+            ${this.computeUom(entity)}
           </span>
           ${isPrimary && this.renderStateTime() || ''}
         </div>
@@ -313,7 +316,7 @@ class MiniGraphCard extends LitElement {
     return html`
       <div class="state__time">
         ${this.tooltip.label ? html`
-          <span>${this.tooltip.label}</span>
+          <span class="tooltip--label">${this.tooltip.label}</span>
         ` : html`
           <span>${this.tooltip.time[0]}</span> -
           <span>${this.tooltip.time[1]}</span>
@@ -323,34 +326,66 @@ class MiniGraphCard extends LitElement {
   }
 
   renderGraph() {
+    const ready = (this.entity[0] && !this.Graph.some(
+      (element, index) => element._history === undefined
+      && this.config.entities[index].show_graph !== false,
+    ))
+    || this.config.show.loading_indicator === false;
     return this.config.show.graph ? html`
       <div class="graph">
-        <div class="graph__container">
-          ${this.renderLabels()}
-          ${this.renderLabelsSecondary()}
-          <div class="graph__container__svg">
-            ${this.renderSvg()}
-          </div>
-        </div>
-        ${this.renderLegend()}
+        ${ready ? html`
+            <div class="graph__container">
+              ${this.renderLabels()}
+              ${this.renderLabelsSecondary()}
+              <div class="graph__container__svg">
+                ${this.renderSvg()}
+              </div>
+            </div>
+            ${this.renderLegend()}
+        ` : html`
+          <ha-circular-progress indeterminate></ha-circular-progress>
+        `}
       </div>` : '';
+  }
+
+  computeLegend(index) {
+    let legend = this.computeName(index);
+    const state = this.getEntityState(index);
+
+    const { show_legend_state = false } = this.config.entities[index];
+
+    if (show_legend_state) {
+      if (this.computeUom(index) === '%') {
+        legend += ` (${this.computeState(state)}${this.computeUom(index)})`;
+      } else {
+        legend += ` (${this.computeState(state)} ${this.computeUom(index)})`;
+      }
+    }
+
+    return legend;
   }
 
   renderLegend() {
     if (this.visibleLegends.length <= 1 || !this.config.show.legend) return;
+
+    /* eslint-disable indent */
     return html`
       <div class="graph__legend">
-        ${this.visibleLegends.map(entity => html`
-          <div class="graph__legend__item"
-            @click=${e => this.handlePopup(e, this.entity[entity.index])}
-            @mouseenter=${() => this.setTooltip(entity.index, -1, this.getEntityState(entity.index), 'Current')}
-            @mouseleave=${() => (this.tooltip = {})}>
-            ${this.renderIndicator(this.getEntityState(entity.index), entity.index)}
-            <span class="ellipsis">${this.computeName(entity.index)}</span>
-          </div>
-        `)}
+        ${this.visibleLegends.map((entity) => {
+          const legend = this.computeLegend(entity.index);
+          return html`
+            <div class="graph__legend__item"
+              @click=${e => this.handlePopup(e, this.entity[entity.index])}
+              @mouseenter=${() => this.setTooltip(entity.index, -1, this.getEntityState(entity.index), 'Current')}
+              @mouseleave=${() => (this.tooltip = {})}>
+              ${this.renderIndicator(this.getEntityState(entity.index), entity.index)}
+              <span class="ellipsis">${legend}</span>
+            </div>
+          `;
+        })}
       </div>
     `;
+    /* eslint-enable indent */
   }
 
   renderIndicator(state, index) {
@@ -656,7 +691,9 @@ class MiniGraphCard extends LitElement {
   }
 
   computeName(index) {
-    return this.config.entities[index].name || this.entity[index].attributes.friendly_name;
+    return this.config.entities[index].name
+      || this.entity[index].attributes.friendly_name
+      || this.entity[index].entity_id;
   }
 
   computeIcon(entity) {
@@ -670,10 +707,17 @@ class MiniGraphCard extends LitElement {
 
   computeUom(index) {
     return (
-      this.config.entities[index].unit
-      || this.config.unit
-      || this.entity[index].attributes.unit_of_measurement
-      || ''
+      this.config.entities[index].unit !== undefined
+        ? this.config.entities[index].unit
+        : (
+          this.config.unit !== undefined
+            ? this.config.unit
+            : (
+              !this.config.entities[index].attribute
+                ? (this.entity[index].attributes.unit_of_measurement || '')
+                : ''
+            )
+        )
     );
   }
 
@@ -795,8 +839,8 @@ class MiniGraphCard extends LitElement {
 
   getBoundaries(series, min, max, fallback, minRange) {
     let boundary = [
-      this.getBoundary('min', series, min, fallback[0], minRange),
-      this.getBoundary('max', series, max, fallback[1], minRange),
+      this.getBoundary('min', series, min, fallback[0]),
+      this.getBoundary('max', series, max, fallback[1]),
     ];
 
     if (minRange) {
@@ -805,10 +849,22 @@ class MiniGraphCard extends LitElement {
 
       // Doesn't matter if minBoundRange is NaN because this will be false if so
       if (diff > 0) {
-        boundary = [
-          boundary[0] - diff / 2,
-          boundary[1] + diff / 2,
+        const weights = [
+          min !== undefined && min[0] !== '~' || max === undefined ? 0 : 1,
+          max !== undefined && max[0] !== '~' || min === undefined ? 0 : 1,
         ];
+        const sum = weights[0] + weights[1];
+        if (sum > 0) {
+          boundary = [
+            boundary[0] - diff * weights[0] / sum,
+            boundary[1] + diff * weights[1] / sum,
+          ];
+        } else {
+          boundary = [
+            boundary[0] - diff / 2,
+            boundary[1] + diff / 2,
+          ];
+        }
       }
     }