+++ src/style.js
@@ -4,7 +4,7 @@ const style = css`
   ha-card {
     flex-direction: column;
     flex: 1;
-    padding: 16px 0;
+    padding: 16px 0 0 0;
     position: relative;
     overflow: hidden;
     height: 100%;
@@ -14,6 +14,10 @@ const style = css`
   }
   ha-card > div:last-child {
     padding-bottom: 0;
+  }
+  ha-card .graph {
+    padding: 0;
+    order: 10;
   }
   ha-card[points] .line--points,
   ha-card[labels] .graph__labels.--primary {
@@ -31,26 +35,17 @@ const style = css`
   ha-card:hover .graph__labels.--secondary {
       opacity: 1;
   }
-  ha-card[fill] {
-    padding-bottom: 0;
-  }
-  ha-card[fill] .graph {
-    padding: 0;
-    order: 10;
-  }
   ha-card[fill] path {
     stroke-linecap: initial;
     stroke-linejoin: initial;
   }
-  ha-card[fill] .graph__legend {
+  ha-card .graph__legend {
     order: -1;
     padding: 0 16px 8px 16px;
   }
-  ha-card[fill] .info {
-    padding-bottom: 16px;
-  }
   ha-card[group] {
     box-shadow: none;
+    border: none;
     padding: 0;
   }
   ha-card[group] > div {
@@ -63,6 +58,9 @@ const style = css`
   }
   ha-card[hover] {
     cursor: pointer;
+  }
+  ha-circular-progress {
+    margin: auto;
   }
   .flex {
     display: flex;